+++ apps/astarte_data_updater_plant/config/config.exs
@@ -23,4 +23,6 @@ config :astarte_data_updater_plant, :amqp_consumer_prefetch_count, 300
 
 config :astarte_rpc, :amqp_queue, "data_updater_plant_rpc"
 
+config :astarte_data_updater_plant, :amqp_adapter, ExRabbitPool.RabbitMQ
+
 import_config "#{config_env()}.exs"